# Optimize DeclarativeEnvironmentRecord