… Jint/Jint.csproj
… Jint/Native/Function/FunctionInstance.cs
… Jint/Native/Function/ScriptFunctionInstance.cs
… Jint/Native/JsNumber.cs
… Jint/Native/JsString.cs
… Jint/Native/Json/JsonSerializer.cs
… Jint/Native/Object/ObjectConstructor.cs
… Jint/Native/Object/ObjectInstance.cs
… Jint/Native/String/StringPrototype.cs
… Jint/Pooling/ArgumentsInstancePool.cs
… Jint/Runtime/Descriptors/PropertyDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/FieldInfoDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/IndexDescriptor.cs
… Jint/Runtime/Descriptors/Specialized/PropertyInfoDescriptor.cs
… Jint/Runtime/Environments/Binding.cs
… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/Environments/EnvironmentRecord.cs
… Jint/Runtime/Environments/LexicalEnvironment.cs
… Jint/Runtime/Environments/ObjectEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/MruPropertyCache.cs
… Jint/Runtime/MruPropertyCache2.cs
… Jint/Runtime/RefStack.cs
… Jint/Runtime/References/Reference.cs
… Jint/Runtime/StatementInterpreter.cs
… Jint/Runtime/StructDictionary.cs
… Jint/Runtime/TypeConverter.cs