* inherit directly from JsValue
* merge MruPropertyCache2 into DeclarativeEnvironmentRecord
* remove MruPropertyCaches as obsolete
* add direct field access to properties that don't inline
* make Binding struct
* use correct overload for GetValue directly
* inline some getter calls
* eliminate some bounds checks
* introduce StructDictionary - a fork from .NET Core Dictionary that specializes for string keys and struct values and ref returns

## ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Slice|100|417.8 us|161.1328|660.16 KB|
| **New** |	|	| **419.3 us (0%)** | **161.1328 (0%)** | **660.16 KB (0%)** |
| Old |Concat|100|468.7 us|175.7813|720.31 KB|
| **New** |	|	| **452.1 us (-4%)** | **175.7813 (0%)** | **720.31 KB (0%)** |
| Old |Unshift|100|18,005.1 us|3562.5000|14672.66 KB|
| **New** |	|	| **18,145.0 us (+1%)** | **3562.5000 (0%)** | **14672.66 KB (0%)** |
| Old |Push|100|10,308.3 us|343.7500|1438.28 KB|
| **New** |	|	| **10,250.2 us (-1%)** | **343.7500 (0%)** | **1438.28 KB (0%)** |
| Old |Index|100|11,806.4 us|390.6250|1637.5 KB|
| **New** |	|	| **11,686.0 us (-1%)** | **390.6250 (0%)** | **1637.5 KB (0%)** |
| Old |Map|100|2,923.6 us|687.5000|2830.47 KB|
| **New** |	|	| **2,588.7 us (-11%)** | **488.2813 (-29%)** | **2008.59 KB (-29%)** |
| Old |Apply|100|599.7 us|188.4766|774.22 KB|
| **New** |	|	| **566.2 us (-6%)** | **188.4766 (0%)** | **774.22 KB (0%)** |
| Old |JsonStringifyParse|100|4,574.5 us|1273.4375|5242.19 KB|
| **New** |	|	| **4,708.5 us (+3%)** | **1273.4375 (0%)** | **5242.19 KB (0%)** |
| Old |FilterWithString|100|32,762.3 us|5687.5000|23335.16 KB|
| **New** |	|	| **28,485.7 us (-13%)** | **3187.5000 (-44%)** | **13179.69 KB (-44%)** |


## ArrayStressBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|560.4 ms|59937.5000|2937.5000|245.9 MB|
| **New** |	|	| **558.0 ms (0%)** | **59937.5000 (0%)** | **2937.5000 (0%)** | **245.9 MB (0%)** |


## DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|73.84 ms|1437.5000|312.5000|-|7722.41 KB|
| **New** |	|	| **69.12 ms (-6%)** | **1312.5000 (-9%)** | **250.0000 (-20%)** | **-** | **7365.02 KB (-5%)** |
| Old |Run|dromaeo-core-eval|15.55 ms|62.5000|-|-|298.71 KB|
| **New** |	|	| **14.89 ms (-4%)** | **62.5000 (0%)** | **-** | **-** | **297.83 KB (0%)** |
| Old |Run|dromaeo-object-array|156.06 ms|42437.5000|1937.5000|937.5000|178313.19 KB|
| **New** |	|	| **156.74 ms (0%)** | **42437.5000 (0%)** | **1937.5000 (0%)** | **937.5000 (0%)** | **178310.42 KB (0%)** |
| Old |Run|dromaeo-object-regexp|883.51 ms|53875.0000|32125.0000|22000.0000|415261.26 KB|
| **New** |	|	| **890.88 ms (+1%)** | **53500.0000 (-1%)** | **31625.0000 (-2%)** | **21500.0000 (-2%)** | **413650.11 KB (0%)** |
| Old |Run|dromaeo-object-string|1,176.09 ms|219000.0000|175687.5000|173125.0000|1393246.66 KB|
| **New** |	|	| **1,175.13 ms (0%)** | **219875.0000 (0%)** | **176687.5000 (+1%)** | **174000.0000 (+1%)** | **1393418.96 KB (0%)** |
| Old |Run|dromaeo-string-base64|168.56 ms|6375.0000|312.5000|-|27381.61 KB|
| **New** |	|	| **163.44 ms (-3%)** | **6375.0000 (0%)** | **250.0000 (-20%)** | **-** | **27380.59 KB (0%)** |


## EvaluationBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Jint|20|667.1 us|118.1641|487.34 KB|
| **New** |	|	| **668.7 us (0%)** | **116.2109 (-2%)** | **477.34 KB (-2%)** |


## LinqJsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|10|70.07 ms|5062.5000|2500.0000|30.52 MB|
| **New** |	|	| **70.04 ms (0%)** | **5062.5000 (0%)** | **2500.0000 (0%)** | **30.52 MB (0%)** |


## MinimalScriptBenchmark

| **Diff**|Method|N|Mean|Gen 0|Allocated|
|------- |-------|-------|-------:|-------:|-------:|
| Old |Jint|10|20.33 us|8.1482|33.44 KB|
| **New** |	|	| **19.76 us (-3%)** | **8.1482 (0%)** | **33.44 KB (0%)** |


## StopwatchBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Jint|1|1.104 s|18000.0000|62.5000|72.22 MB|
| **New** |	|	| **1.092 s (-1%)** | **12562.5000 (-30%)** | **62.5000 (0%)** | **50.49 MB (-30%)** |


## SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|579.8 ms|11875.0000|312.5000|-|49.56 MB|
| **New** |	|	| **551.3 ms (-5%)** | **11375.0000 (-4%)** | **312.5000 (0%)** | **-** | **47.56 MB (-4%)** |
| Old |Run|3d-morph|563.1 ms|11812.5000|2500.0000|1062.5000|68.03 MB|
| **New** |	|	| **548.2 ms (-3%)** | **11812.5000 (0%)** | **2500.0000 (0%)** | **1062.5000 (0%)** | **68.03 MB (0%)** |
| Old |Run|3d-raytrace|465.7 ms|21875.0000|750.0000|62.5000|89.75 MB|
| **New** |	|	| **439.8 ms (-6%)** | **19000.0000 (-13%)** | **875.0000 (+17%)** | **125.0000 (+100%)** | **78.4 MB (-13%)** |
| Old |Run|access-binary-trees|191.2 ms|16625.0000|250.0000|-|68.51 MB|
| **New** |	|	| **189.4 ms (-1%)** | **13000.0000 (-22%)** | **375.0000 (+50%)** | **-** | **54.07 MB (-21%)** |
| Old |Run|access-fannkuch|1,831.1 ms|9437.5000|125.0000|-|37.87 MB|
| **New** |	|	| **1,739.4 ms (-5%)** | **9437.5000 (0%)** | **125.0000 (0%)** | **-** | **37.87 MB (0%)** |
| Old |Run|access-nbody|551.5 ms|13062.5000|-|-|52.42 MB|
| **New** |	|	| **659.3 ms (+20%)** | **12812.5000 (-2%)** | **62.5000** | **-** | **51.31 MB (-2%)** |
| Old |Run|access-nsieve|712.1 ms|15375.0000|3937.5000|1875.0000|73.09 MB|
| **New** |	|	| **655.2 ms (-8%)** | **15375.0000 (0%)** | **3937.5000 (0%)** | **1875.0000 (0%)** | **73.09 MB (0%)** |
| Old |Run|bitops-3bit-bits-in-byte|383.9 ms|17562.5000|-|-|70.27 MB|
| **New** |	|	| **401.1 ms (+4%)** | **12625.0000 (-28%)** | **-** | **-** | **50.74 MB (-28%)** |
| Old |Run|bitops-bits-in-byte|647.5 ms|11562.5000|62.5000|-|46.44 MB|
| **New** |	|	| **590.5 ms (-9%)** | **8187.5000 (-29%)** | **-** | **-** | **32.77 MB (-29%)** |
| Old |Run|bitops-bitwise-and|476.8 ms|9000.0000|-|-|36.01 MB|
| **New** |	|	| **487.9 ms (+2%)** | **9000.0000 (0%)** | **-** | **-** | **36.01 MB (0%)** |
| Old |Run|bitops-nsieve-bits|685.8 ms|21625.0000|125.0000|-|88.5 MB|
| **New** |	|	| **678.7 ms (-1%)** | **21625.0000 (0%)** | **125.0000 (0%)** | **-** | **88.5 MB (0%)** |
| Old |Run|controlflow-recursive|265.5 ms|23625.0000|-|-|94.61 MB|
| **New** |	|	| **236.8 ms (-11%)** | **16187.5000 (-31%)** | **-** | **-** | **64.95 MB (-31%)** |
| Old |Run|crypto-aes|430.7 ms|5500.0000|312.5000|-|24.12 MB|
| **New** |	|	| **411.0 ms (-5%)** | **5062.5000 (-8%)** | **250.0000 (-20%)** | **-** | **22.47 MB (-7%)** |
| Old |Run|crypto-md5|301.7 ms|28000.0000|250.0000|62.5000|114.57 MB|
| **New** |	|	| **278.0 ms (-8%)** | **23875.0000 (-15%)** | **250.0000 (0%)** | **62.5000 (0%)** | **97.73 MB (-15%)** |
| Old |Run|crypto-sha1|292.6 ms|22687.5000|312.5000|62.5000|92.28 MB|
| **New** |	|	| **320.4 ms (+10%)** | **19125.0000 (-16%)** | **250.0000 (-20%)** | **-** | **78.4 MB (-15%)** |
| Old |Run|date-format-tofte|318.2 ms|11687.5000|62.5000|-|47.06 MB|
| **New** |	|	| **303.6 ms (-5%)** | **11187.5000 (-4%)** | **-** | **-** | **45 MB (-4%)** |
| Old |Run|date-format-xparb|322.6 ms|10125.0000|250.0000|-|41.22 MB|
| **New** |	|	| **313.1 ms (-3%)** | **9062.5000 (-10%)** | **250.0000 (0%)** | **-** | **37.09 MB (-10%)** |
| Old |Run|math-cordic|910.4 ms|20562.5000|62.5000|-|82.43 MB|
| **New** |	|	| **843.3 ms (-7%)** | **16687.5000 (-19%)** | **62.5000 (0%)** | **-** | **66.98 MB (-19%)** |
| Old |Run|math-partial-sums|279.4 ms|7000.0000|-|-|28.07 MB|
| **New** |	|	| **276.7 ms (-1%)** | **7000.0000 (0%)** | **-** | **-** | **28.06 MB (0%)** |
| Old |Run|math-spectral-norm|323.3 ms|16875.0000|-|-|67.63 MB|
| **New** |	|	| **302.6 ms (-6%)** | **13125.0000 (-22%)** | **-** | **-** | **52.65 MB (-22%)** |
| Old |Run|regexp-dna|342.7 ms|2250.0000|1875.0000|1562.5000|13.54 MB|
| **New** |	|	| **341.1 ms (0%)** | **2250.0000 (0%)** | **1875.0000 (0%)** | **1562.5000 (0%)** | **13.54 MB (0%)** |
| Old |Run|string-base64|252.4 ms|9312.5000|375.0000|-|38.79 MB|
| **New** |	|	| **245.6 ms (-3%)** | **9312.5000 (0%)** | **375.0000 (0%)** | **-** | **38.79 MB (0%)** |
| Old |Run|string-fasta|417.4 ms|26000.0000|62.5000|-|104.02 MB|
| **New** |	|	| **391.3 ms (-6%)** | **24562.5000 (-6%)** | **62.5000 (0%)** | **-** | **98.46 MB (-5%)** |
| Old |Run|string-tagcloud|200.6 ms|14125.0000|2062.5000|750.0000|64.54 MB|
| **New** |	|	| **181.8 ms (-9%)** | **12625.0000 (-11%)** | **1750.0000 (-15%)** | **750.0000 (0%)** | **58.98 MB (-9%)** |
| Old |Run|string-unpack-code|157.2 ms|13687.5000|4375.0000|937.5000|64.88 MB|
| **New** |	|	| **149.3 ms (-5%)** | **11000.0000 (-20%)** | **3375.0000 (-23%)** | **875.0000 (-7%)** | **54.79 MB (-16%)** |
| Old |Run|string-validate-input|174.2 ms|7250.0000|437.5000|62.5000|33.26 MB|
| **New** |	|	| **164.7 ms (-5%)** | **6750.0000 (-7%)** | **375.0000 (-14%)** | **62.5000 (0%)** | **31.03 MB (-7%)** |


## UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|313.6 ms|46750.0000|250.0000|187.34 MB|
| **New** |	|	| **293.5 ms (-6%)** | **36812.5000 (-21%)** | **125.0000 (-50%)** | **147.43 MB (-21%)** |